app/auto/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
allure-pytest==2.9.45
allure-python-commons==2.9.45
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.79.1
frozenlist==1.3.1
greenlet==1.1.3
h11==0.13.0
idna==3.3
loguru==0.6.0
multidict==6.0.2
pool==0.1.2.dev0
pydantic==1.9.2
pyinstrument==4.3.0
pytest-assume==2.4.3
PyYAML==6.0
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
sniffio==1.2.0
SQLAlchemy==1.4.41
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.2
yarl==1.8.1
